requirements.txt
asgiref==3.7.2
cffi==1.16.0
crispy-bootstrap5==2023.10
cryptography==41.0.7
Django==5.0
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-js-asset==2.2.0
django-taggit==5.0.1
django-taggit-serializer==0.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
fuzzywuzzy==0.18.0
Levenshtein==0.23.0
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.0
python-Levenshtein==0.23.0
pytz==2023.3.post1
rapidfuzz==3.6.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
whitenoise==6.6.0